+++ src/mui/list/Datagrid.js
@@ -1,8 +1,6 @@
 import React, { Component, PropTypes } from 'react';
-import { TableHeaderColumn, TableRowColumn } from 'material-ui/Table';
-import FlatButton from 'material-ui/FlatButton';
-import ContentSort from 'material-ui/svg-icons/content/sort';
-import title from '../../util/title';
+import DatagridCell from './DatagridCell';
+import DatagridHeaderCell from './DatagridHeaderCell';
 
 const defaultStyles = {
     table: {
@@ -20,23 +18,70 @@ const defaultStyles = {
         color: 'rgba(0, 0, 0, 0.870588)',
         height: 48,
     },
-    'th:first-child': {
-        padding: '0 0 0 12px',
+    header: {
+        th: {
+            padding: 0,
+        },
+        'th:first-child': {
+            padding: '0 0 0 12px',
+        },
+        sortButton: {
+            minWidth: 40,
+        },
+        nonSortableLabel: {
+            position: 'relative',
+            paddingLeft: 16,
+            paddingRight: 16,
+            verticalAlign: 'middle',
+            letterSpacing: 0,
+            textTransform: 'uppercase',
+            fontWeight: 500,
+            fontSize: 14,
+        },
     },
-    th: {
-        padding: 0,
-    },
-    sortButton: {
-        minWidth: 40,
-    },
-    'td:first-child': {
-        padding: '0 12px 0 24px',
-    },
-    td: {
-        padding: '0 12px',
+    cell: {
+        td: {
+            padding: '0 12px',
+        },
+        'td:first-child': {
+            padding: '0 12px 0 24px',
+        },
     },
 };
 
+/**
+ * The Datagrid component renders a list of records as a table.
+ * It is usually used as a child of the <List> and <ReferenceManyField> components.
+ *
+ * Props:
+ *  - styles
+ *  - rowStyle
+ *
+ * @example Display all posts as a datagrid
+ * const postRowStyle = (record, index) => ({
+ *     backgroundColor: record.nb_views >= 500 ? '#efe' : 'white',
+ * });
+ * export const PostList = (props) => (
+ *     <List {...props}>
+ *         <Datagrid rowStyle={postRowStyle}>
+ *             <TextField source="id" />
+ *             <TextField source="title" />
+ *             <TextField source="body" />
+ *             <EditButton />
+ *         </Datagrid>
+ *     </List>
+ * );
+ *
+ * @example Display all the comments of the current post as a datagrid
+ * <ReferenceManyField reference="comments" target="post_id">
+ *     <Datagrid>
+ *         <TextField source="id" />
+ *         <TextField source="body" />
+ *         <DateField source="created_at" />
+ *         <EditButton />
+ *     </Datagrid>
+ * </ReferenceManyField>
+ */
 class Datagrid extends Component {
     updateSort = (event) => {
         event.stopPropagation();
@@ -44,36 +89,34 @@ class Datagrid extends Component {
     }
 
     render() {
-        const { resource, children, ids, data, currentSort, basePath, styles = defaultStyles, updateSort } = this.props;
+        const { resource, children, ids, data, currentSort, basePath, styles = defaultStyles, rowStyle, updateSort } = this.props;
         return (
             <table style={styles.table}>
                 <thead>
                     <tr style={styles.tr}>
                         {React.Children.map(children, (field, index) => (
-                            <TableHeaderColumn key={field.props.source || index} style={index ? styles.th : styles['th:first-child']} >
-                                {(field.props.label || field.props.source) &&
-                                    <FlatButton
-                                        labelPosition="before"
-                                        onClick={this.updateSort}
-                                        data-sort={field.props.source}
-                                        label={title(field.props.label, field.props.source)}
-                                        icon={field.props.source === currentSort.sort ?
-                                            <ContentSort style={currentSort.order === 'ASC' ? { transform: 'rotate(180deg)' } : {}} /> : false
-                                        }
-                                        style={styles.sortButton}
-                                    />
-                                }
-                            </TableHeaderColumn>
+                            <DatagridHeaderCell
+                                key={field.props.source || index}
+                                field={field}
+                                defaultStyle={styles.header}
+                                isFirst={index === 0}
+                                currentSort={currentSort}
+                                updateSort={this.updateSort}
+                            />
                         ))}
                     </tr>
                 </thead>
                 <tbody style={styles.tbody}>
-                    {ids.map(id => (
-                        <tr style={styles.tr} key={id}>
+                    {ids.map((id, index) => (
+                        <tr style={rowStyle ? rowStyle(data[id], index) : styles.tr} key={id}>
                             {React.Children.toArray(children).map((field, index) => (
-                                <TableRowColumn key={`${id}-${field.props.source || index}`} style={index ? styles.td : styles['td:first-child']} >
-                                    <field.type {...field.props} record={data[id]} basePath={basePath} resource={resource} />
-                                </TableRowColumn>
+                                <DatagridCell
+                                    key={`${id}-${field.props.source || index}`}
+                                    record={data[id]}
+                                    defaultStyle={styles.cell}
+                                    isFirst={index === 0}
+                                    {...{ field, basePath, resource }}
+                                />
                             ))}
                         </tr>
                     ))}
@@ -94,6 +137,7 @@ Datagrid.propTypes = {
     basePath: PropTypes.string,
     setSort: PropTypes.func,
     styles: PropTypes.object,
+    rowStyle: PropTypes.func,
 };
 
 Datagrid.defaultProps = {